pyproject.toml
[project]
name = "a2a-proxy"
version = "0.1.0"
description = "Agent-to-Agent minimal proxy for Elastic endpoints (Cloud Run ready)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "A2A" }]
license = { text = "Apache-2.0" }
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "orjson>=3.10.0"
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[tool.ruff]
line-length = 100

[tool.pyright]
include = ["app"]
